# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fee Manager API",
    "description": "REST API for managing validator configurations for Vouch and Commit-Boost",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "{server_url}",
      "variables": {
        "server_url": {
          "default": "http://localhost:3000",
          "description": "API Server URL"
        }
      }
    }
  ],
  "paths": {
...
    "/api/admin/vouch/proposer-patterns": {
      "get": {
        "tags": [
          "Vouch - Proposer Patterns"
        ],
        "operationId": "list_proposer_patterns",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pattern",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fee_recipient",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "gas_limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "min_value",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reset_relays",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "relay_url",
            "in": "query",
            "description": "Filter by relay URL (prefix match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "relay_min_value",
            "in": "query",
            "description": "Filter by relay min_value (exact match)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "relay_disabled",
            "in": "query",
            "description": "Filter by relay disabled status",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return (1-1000, default 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "after",
            "in": "query",
            "description": "Opaque cursor: the `next_cursor` of the previous page",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of proposer patterns",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_ProposerPatternListItem"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Vouch - Proposer Patterns"
        ],
        "operationId": "create_proposer_pattern",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProposerPatternRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Pattern created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposerPatternResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid field value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Pattern already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/admin/vouch/proposer-patterns/{name}": {
      "get": {
        "tags": [
          "Vouch - Proposer Patterns"
        ],
        "operationId": "get_proposer_pattern",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Pattern name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposer pattern details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposerPatternResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pattern not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Vouch - Proposer Patterns"
        ],
        "operationId": "update_proposer_pattern",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Pattern name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProposerPatternRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Pattern updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposerPatternResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid field value",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pattern not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Vouch - Proposer Patterns"
        ],
        "operationId": "delete_proposer_pattern",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Pattern name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Pattern deleted"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Pattern not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProposerPatternRequest": {
        "type": "object",
        "required": [
          "name",
          "pattern"
        ],
        "properties": {
          "fee_recipient": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/String"
              },
              {
                "type": "null"
              }
            ]
          },
          "gas_limit": {
            "type": [
              "string",
              "null"
            ]
          },
          "min_value": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "pattern": {
            "type": "string"
          },
          "relays": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/ProposerRelayConfig"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "reset_relays": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "ListResponse_ProposerPatternListItem": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "name",
                "pattern",
                "tags",
                "reset_relays",
                "created_at",
                "updated_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "fee_recipient": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/String"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "gas_limit": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "min_value": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "name": {
                  "type": "string"
                },
                "pattern": {
                  "type": "string"
                },
                "reset_relays": {
                  "type": "boolean"
                },
                "tags": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Pass as `after` to get the next page; absent on the last page"
          }
        }
      },
...
      "ProposerPatternResponse": {
        "type": "object",
        "required": [
          "name",
          "pattern",
          "tags",
          "reset_relays",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "fee_recipient": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/String"
              },
              {
                "type": "null"
              }
            ]
          },
          "gas_limit": {
            "type": [
              "string",
              "null"
            ]
          },
          "min_value": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "pattern": {
            "type": "string"
          },
          "relays": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/ProposerRelayConfig"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "reset_relays": {
            "type": "boolean"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "String": {
        "type": "array",
        "items": {
          "type": "integer",
          "format": "int32",
          "minimum": 0
        },
        "description": "BLS public key (48 bytes, hex-encoded with 0x prefix)",
        "example": "0x8021cef5a095e4a4476d5a56d0e63e7ddb5e66ba5349fe1bc341f94f0f07a0dbfd19755b9820022e6b9153e28fce90e8"
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token",
        "description": "API token for authentication"
      }
    }
  },
...
}